+++ lib/notifications/windows.dart
@@ -3,7 +3,7 @@ import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 const WindowsInitializationSettings initializationSettings =
     WindowsInitializationSettings(
   appName: 'Letscheck',
-  appUserModelId: 'dev.jochum.letscheck',
+  appUserModelId: 'io.github.jochumdev.letscheck',
   guid: '7dba2765-0e22-4048-b0f9-80ab10eba346',
   iconPath: 'assets/icons/letscheck.ico',
 );